Update Work Experiences
I updated the Outside the Box page (should be finished, for now)
I created and filled in company descriptions for the InMoment / MaritzCX page
I updated the main page to have better wording
I made some small adjustments to the work experience page, and changed the file name to be "work-experience" for better clarity.  I updated the menu to match.